Implement listpids() for Linux
